Exclude trailing parenthesis from URL
I found that when a link was enclosed in parentheses, e.g., (https://google.com), the opening parenthesis was correctly excluded from the URL, but the closing one was included. This is fixed here using the same pattern that excludes closing periods, but there may be a better way (either earlier in the regex, or maybe the exclusions should be in a foreach loop).